enormes changements

+++ park-core/src/main/java/junia/park/core/dao/LocalisationDAO.java
@@ -3,6 +3,6 @@ package junia.park.core.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import junia.park.core.entity.Localisation;
 
-public interface LocalisationDAO extends JpaRepository<Localisation,Long> {
+public interface LocalisationDAO extends JpaRepository<Localisation,Integer> {
 
 }
