+++ park-core/src/main/java/junia/park/core/dao/ParkingDAO.java
@@ -3,6 +3,6 @@ package junia.park.core.dao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import junia.park.core.entity.Parking;
 
-public interface ParkingDAO extends JpaRepository<Parking, Long> {
+public interface ParkingDAO extends JpaRepository<Parking, Integer> {
 
 }
